Move files from tools/init_cfg to tools preserving compatibility
