+++ src/tools/__init__.py
@@ -18,9 +18,9 @@
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <http://www.gnu.org/licenses/>. 
 
-
+# compatibility with previous ESPP versions:
 import espressopp.tools.convert
-from espressopp.tools.init_cfg import *
+import espressopp.tools.init_cfg
 
 from espressopp.tools.analyse import *
 from espressopp.tools.decomp import *
@@ -30,6 +30,7 @@ from espressopp.tools.gromacs import *
 from espressopp.tools.info import *
 from espressopp.tools.io_extended import *
 from espressopp.tools.lammps import *
+from espressopp.tools.lattice import *
 from espressopp.tools.pathintegral import *
 from espressopp.tools.pdb import *
 from espressopp.tools.povwrite import *
@@ -42,5 +43,6 @@ from espressopp.tools.timers import *
 from espressopp.tools.topology import *
 from espressopp.tools.topology_helper import *
 from espressopp.tools.units import *
+from espressopp.tools.velocities import *
 from espressopp.tools.vmd import *
 from espressopp.tools.warmup import *
